+++ src/ContaoManager/Plugin.php
@@ -29,7 +29,7 @@ class Plugin implements BundlePluginInterface, ConfigPluginInterface
      *
      * @return ConfigInterface[]
      */
-    public function getBundles(ParserInterface $parser)
+    public function getBundles(ParserInterface $parser): array
     {
         return [
           BundleConfig::create(HeimrichHannotSearchBundle::class)->setLoadAfter([